修改吞吐率为1秒计算一次

+++ easycluster-cluster/src/main/java/org/easycluster/easycluster/cluster/netty/NetworkClientStatisticsMBean.java
@@ -1,7 +1,7 @@
 package org.easycluster.easycluster.cluster.netty;
 
 public interface NetworkClientStatisticsMBean {
-	int getRequestsPerSecond();
+	double getRequestsPerSecond();
 
-	long getAverageRequestProcessingTime();
+	double getAverageRequestProcessingTime();
 }
